feat: restructure input components by moving email and password inputs to a new directory and updating their references
