+++ src/app/shared/components/input-elements/password-input/password-input.component.scss
